updated id data type

+++ src/test/scala/org/thompson/ingestion/model/Employee.scala
@@ -1,5 +1,5 @@
 package org.thompson.ingestion.model
 
-final case class Employee(id: Int, name: String, department: String)
+final case class Employee(id: Long, name: String, department: String)
 
-final case class NewEmployee(id: Int, name: String, department: String, mobile: String)
+final case class NewEmployee(id: Long, name: String, department: String, mobile: String)
